exc 4.21*

+++ part4/blog-backend/controllers/blogs.js
@@ -3,13 +3,8 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
+
+
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -23,11 +18,10 @@ blogsRouter.get('/:id', async (request, response, next) => {
 })
   
 blogsRouter.post('/', async (request, response, next) => {
-    const body = request.body      
-    const token = getTokenFrom(request)
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
+    if (!request.token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
     }
     const user = await User.findById(decodedToken.id)
@@ -47,8 +41,16 @@ blogsRouter.post('/', async (request, response, next) => {
   
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  const blog = await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()   
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  
+
+  if ( decodedToken.id === blog.user.toString() ) {
+    const blog = await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end() 
+  }
+  return response.status(401).json({ error: 'you do not own the blog entry' })
+
   })
 
 blogsRouter.put('/:id', async (request, response, next) => {
